Fixed waitTime variable type, corrected wait time string name in smartdashboard

+++ 2021/OffSeason2021/src/main/include/commands/AutoDriveShoot.h
@@ -42,5 +42,5 @@ public:
 
 private:
     const char *path;
-    int waitTime;
+    double waitTime;
 };
